Refactor code

+++ LockerService.Application/EventBus/Mqtt/Events/MqttUpdateLockerStatusEvent.cs
@@ -2,11 +2,11 @@ namespace LockerService.Application.EventBus.Mqtt.Events;
 
 public class MqttUpdateLockerStatusEvent : MqttBaseMessage
 {
-    public int LockerId { get; private set; }
+    public long LockerId { get; private set; }
  
     public string Status { get; private set; }
 
-    public MqttUpdateLockerStatusEvent(int lockerId, LockerStatus status) : base()
+    public MqttUpdateLockerStatusEvent(long lockerId, LockerStatus status) : base()
     {
         LockerId = lockerId;
         Status = status.ToString();
